handle withdraw and deposit account

+++ auth/repository.go
@@ -6,6 +6,6 @@ import (
 
 type UserRepository interface {
 	CreateUser(ctx context.Context, username string, password string) error
-	GetUser(ctx context.Context, id int) (*User, error)
+	GetUser(ctx context.Context, id int64) (*User, error)
 	GetUserByUsername(ctx context.Context, username string) (*User, error)
 }
